// packages/datadog-api-client-v1/models/SyntheticsAssertionJSONPathTargetTarget.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Composed target for `validatesJSONPath` operator.
*/
export declare class SyntheticsAssertionJSONPathTargetTarget {
    /**
    * The JSON path to assert.
    */
    'jsonPath'?: string;
    /**
    * The specific operator to use on the path.
    */
    'operator'?: string;
    /**
    * The path target value to compare to.
    */
    'targetValue'?: any;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SyntheticsAssertionJSONPathTargetTarget;
    static serialize(data: SyntheticsAssertionJSONPathTargetTarget): {
        [key: string]: any;
    };
    constructor();
}